requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
pillow==10.1.0
requests==2.31.0
sqlalchemy==2.0.45
typing-extensions==4.15.0
urllib3==2.6.2
werkzeug==3.0.1
